Change timing profile variable to use unsigned long long instead of double.

+++ hplayer/source/include/cgmiDiagsApi.h
@@ -56,7 +56,7 @@ typedef struct
     tCgmiDiag_timingEvent timingEvent;
     char sessionUri[1024];
     unsigned int sessionIndex;
-    double markTime;
+    unsigned long long markTime;
 }tCgmiDiags_timingMetric;
